Finished adsDaoTest
