+++ notes
@@ -1,7 +1,5 @@
 - Show users ads on their profile page
 
-- Add buttons to allow users to update and delete ads
-
 - Allow users to update their profile information
 
 - validate data that goes to database
