--- src/main/java/com/codeup/adlister/dao/MySQLAdsDaoTest.java
@@ -1,36 +1,0 @@
-package com.codeup.adlister.dao;
-
-import org.junit.jupiter.api.Test;
-import org.mockito.Mock;
-
-import java.util.HashMap;
-import java.util.List;
-
-import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.Mockito.*;
-
-class MySQLAdsDaoTest {
-
-    @Mock
-    List<Ads> MockList;
-
-    @Test
-    void all() {
-    }
-
-    @Test
-    void insert() {
-    }
-
-    @Test
-    void findAdById() {
-    }
-
-    @Test
-    void adsTitleQuery() {
-    }
-
-    @Test
-    void delete() {
-    }
-}
